# backend/src/api/openapi.yaml
openapi: 3.1.0
info:
  title: PMP Learning Management API
  description: Comprehensive API for PMBOK 6th and 7th Edition learning management system
  version: 1.0.0
  contact:
    name: API Support
    email: api@pmplm.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.pmplm.com/v1
    description: Production server
...
paths:
...
  /principles:
    get:
      tags: [Principles]
      summary: List all principles
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SearchParam'
        - name: complexity
          in: query
          schema:
            type: string
            enum: [LOW, MEDIUM, HIGH, VERY_HIGH]
        - name: includeActions
          in: query
          schema:
            type: boolean
            default: false
      responses:
        200:
          description: List of principles
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Principle'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /principles/{id}:
    get:
      tags: [Principles]
      summary: Get principle by ID
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - name: includeAll
          in: query
          schema:
            type: boolean
            default: false
      responses:
        200:
          description: Principle details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/PrincipleDetail'
        404:
          $ref: '#/components/responses/NotFound'
  /principles/{id}/relationships:
    get:
      tags: [Principles]
      summary: Get principle relationships
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        200:
          description: Principle relationships
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      domains:
                        type: array
                        items:
                          $ref: '#/components/schemas/DomainMapping'
                      resources:
                        type: array
                        items:
                          $ref: '#/components/schemas/Resource'

  # Performance Domains endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: cuid
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SearchParam:
      name: search
      in: query
      schema:
        type: string
        minLength: 2
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
...
    Principle:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        nameEn:
          type: string
        description:
          type: string
        complexity:
          type: string
        order:
          type: integer
    PrincipleDetail:
      allOf:
        - $ref: '#/components/schemas/Principle'
        - type: object
          properties:
            details:
              type: string
            keyActions:
              type: array
              items:
                $ref: '#/components/schemas/KeyAction'
            examples:
              type: array
              items:
                $ref: '#/components/schemas/PrincipleExample'
            domainMappings:
              type: array
              items:
                $ref: '#/components/schemas/DomainMapping'
    KeyAction:
      type: object
      properties:
        id:
          type: string
        action:
          type: string
        actionEn:
          type: string
        description:
          type: string
        order:
          type: integer
    PrincipleExample:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        scenario:
          type: string
        application:
          type: string
        outcome:
          type: string
        industry:
          type: string
        projectType:
          type: string
    PerformanceDomain:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        nameEn:
          type: string
        description:
          type: string
        complexity:
          type: string
        order:
          type: integer
...
    DomainMapping:
      type: object
      properties:
        id:
          type: string
        domain:
          $ref: '#/components/schemas/PerformanceDomain'
        mappingType:
          type: string
          enum: [PRIMARY, SECONDARY, SUPPORTING, RELATED]
        relevanceScore:
          type: integer
          minimum: 1
          maximum: 10
        description:
          type: string
...
    Resource:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [VIDEO, ARTICLE, EBOOK, PODCAST, INFOGRAPHIC, TEMPLATE, CASE_STUDY, SIMULATION]
        url:
          type: string
        duration:
          type: integer
        difficulty:
          type: string
        isPremium:
          type: boolean
...